feat!: rename plugin to 'antonyms'
