+++ dev/init.lua
@@ -1,11 +1,11 @@
 -- force lua to import the modules again
 package.loaded["dev"] = nil
-package.loaded["alterego"] = nil
+package.loaded["antonyms"] = nil
 
 vim.api.nvim_set_keymap("n", ",r", "<cmd>luafile dev/init.lua<cr>", {})
 
-require("alterego").setup({
-    alteregos = {
+require("antonyms").setup({
+    antonyms = {
         { "bleu", "rouge" }
     }
 })
